pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["LevBratisman <leobratisman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.6.4"}

[tool.poetry.scripts]
serve = "app.main:run"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
